Class to store and retrieve GREAT results

Class to store and retrieve GREAT results

Details

After submitting request to GREAT server, the generated results will be available on GREAT server for some time. The GreatJob-class is defined to store parameters that user has set and result tables what were retrieved from GREAT server.

Constructor

Users don't need to construct by hand, submitGreatJob is used to generate a GreatJob-class instance.

Workflow

After submitting request to GREAT server, users can perform following steps:

getEnrichmentTables,GreatJob-method to get enrichment tables for selected ontologies catalogues.

plotRegionGeneAssociations,GreatJob-method to plot associations between regions and genes

getRegionGeneAssociations,GreatJob-method to get a GRanges object which contains associations bewteen regions and genes.

shinyReport,GreatJob-method to view the results by a shiny application.

Author

Zuguang gu <z.gu@dkfz.de>

Examples

# There is no example
NULL
#> NULL