Get a single enrichment table from GREAT web server

Get a single enrichment table from GREAT web server

# S4 method for GreatJob
getEnrichmentTable(object, ontology, ...)

Arguments

object

A GreatJob-class object returned by submitGreatJob.

ontology

A single ontology names. Valid values are in availableOntologies.

...

All pass to getEnrichmentTables,GreatJob-method.

Value

A data frame of the enrichment results for a single ontology.

Examples

# note the `job` was generated from GREAT 3.0.0
job = readRDS(system.file("extdata", "job.rds", package = "rGREAT"))
tb = getEnrichmentTable(job, ontology = "GO Molecular Function")
#> The default enrichment tables contain no associated genes for the input
#> regions.You can set `download_by = 'tsv'` to download the complete
#> table,but note only the top 500 regions can be retreived. See the
#> following link:
#> 
#> https://great-help.atlassian.net/wiki/spaces/GREAT/pages/655401/Export#Export-GlobalExport
head(tb)
#>           ID
#> 1 GO:0070002
#> 2 GO:0004731
#> 3 GO:0016841
#> 4 GO:0016901
#> 5 GO:0008111
#> 6 GO:0003796
#>                                                                                                    name
#> 1                                                                      glutamic-type peptidase activity
#> 2                                                              purine-nucleoside phosphorylase activity
#> 3                                                                                ammonia-lyase activity
#> 4 oxidoreductase activity, acting on the CH-OH group of donors, quinone or similar compound as acceptor
#> 5                                                                alpha-methylacyl-CoA racemase activity
#> 6                                                                                     lysozyme activity
#>   Binom_Genome_Fraction Binom_Expected Binom_Observed_Region_Hits
#> 1          8.106316e-06    0.008146847                          1
#> 2          1.435477e-05    0.014426540                          1
#> 3          1.864956e-04    0.187428100                          2
#> 4          9.768315e-04    0.981715700                          4
#> 5          1.855838e-05    0.018651170                          1
#> 6          9.926527e-04    0.997616000                          4
#>   Binom_Fold_Enrichment Binom_Region_Set_Coverage Binom_Raw_PValue
#> 1            122.746900              0.0009950249      0.008113784
#> 2             69.316690              0.0009950249      0.014323080
#> 3             10.670760              0.0019900500      0.015504970
#> 4              4.074500              0.0039801000      0.017829270
#> 5             53.615940              0.0009950249      0.018478480
#> 6              4.009559              0.0039801000      0.018781680
#>   Binom_Adjp_BH Hyper_Total_Genes Hyper_Expected Hyper_Observed_Gene_Hits
#> 1             1                 1     0.08785544                        1
#> 2             1                 1     0.08785544                        1
#> 3             1                 5     0.43927720                        2
#> 4             1                 5     0.43927720                        3
#> 5             1                 1     0.08785544                        1
#> 6             1                11     0.96640980                        4
#>   Hyper_Fold_Enrichment Hyper_Gene_Set_Coverage Hyper_Term_Gene_Coverage
#> 1             11.382330            0.0006309148                1.0000000
#> 2             11.382330            0.0006309148                1.0000000
#> 3              4.552934            0.0012618300                0.4000000
#> 4              6.829401            0.0018927440                0.6000000
#> 5             11.382330            0.0006309148                1.0000000
#> 6              4.139031            0.0025236590                0.3636364
#>   Hyper_Raw_PValue Hyper_Adjp_BH
#> 1      0.087855440     1.0000000
#> 2      0.087855440     1.0000000
#> 3      0.064472140     1.0000000
#> 4      0.005910147     0.4590875
#> 5      0.087855440     1.0000000
#> 6      0.011837350     0.5820820