Get region-gene associations

Get region-gene associations

# S4 method for GreatJob
getRegionGeneAssociations(object, ontology = NULL, term_id = NULL,
    request_interval = 10, max_tries = 100, verbose = great_opt$verbose)

Arguments

object

A GreatJob-class object returned by submitGreatJob.

ontology

ontology name

term_id

Term id in the selected ontology.

request_interval

Time interval for two requests. Default is 300 seconds.

max_tries

Maximal times for automatically reconnecting GREAT web server.

verbose

Whether to show messages.

Value

A GRanges object. Please the two meta columns are in formats of CharacterListand IntegerList because a region may associate to multiple genes. Please note, the distance is from the middle points of input regions to TSS.

Author

Zuguang gu <z.gu@dkfz.de>

Examples

job = readRDS(system.file("extdata", "GreatJob.rds", package = "rGREAT"))

gr = getRegionGeneAssociations(job)
#> Warning: EOF within quoted string
#> Error in if (data[i, 2] == "NONE") {    region = c(region, data[i, 1])    gene = c(gene, NA)    distance = c(distance, NA)} else {    r = gregexpr("([a-zA-Z0-9\\-_.]+) \\(([+-]?\\d+)\\)", data[i,         2], perl = TRUE)[[1]]    capture.start = attr(r, "capture.start")    capture.length = attr(r, "capture.length")    k = nrow(capture.start)    region = c(region, rep(data[i, 1], k))    gene = c(gene, substr(rep(data[i, 2], k), capture.start[,         1], capture.start[, 1] + capture.length[, 1] - 1))    distance = c(distance, substr(rep(data[i, 2], k), capture.start[,         2], capture.start[, 2] + capture.length[, 2] - 1))}: argument is of length zero
gr
#> Error in eval(expr, envir, enclos): object 'gr' not found