Get region-gene associations

Get region-gene associations

# S4 method for GreatObject
getRegionGeneAssociations(object, term_id = NULL, by_middle_points = FALSE,
    use_symbols = TRUE)

Arguments

object

A GreatObject-class object returned by great.

term_id

Term ID.

by_middle_points

Whether the distances are calculated from the middle points of input regions?

use_symbols

Whether to use gene symbols

Value

A GRanges object. Please the two meta columns are in formats of CharacterListand IntegerList because a region may associate to multiple genes.

Examples

obj = readRDS(system.file("extdata", "GreatObject.rds", package = "rGREAT"))
getRegionGeneAssociations(obj)
#> GRanges object with 884 ranges and 2 metadata columns:
#>         seqnames            ranges strand | annotated_genes    dist_to_TSS
#>            <Rle>         <IRanges>  <Rle> | <CharacterList>  <IntegerList>
#>     [1]     chr1   9204434-9208784      * |     GPR157,H6PD  -15205,-86079
#>     [2]     chr1   9853594-9859363      * |   PIK3CD,CLSTN1   141804,25187
#>     [3]     chr1 10862809-10871681      * |  CASZ1,C1orf127   -6076,170413
#>     [4]     chr1 12716970-12723206      * | AADACL4,AADACL3   12404,-52912
#>     [5]     chr1 13814692-13823250      * | PRAMEF17,LRRC38    98604,16992
#>     ...      ...               ...    ... .             ...            ...
#>   [880]     chrY 16384799-16389852      * |   TMSB4Y,NLGN4Y 569352,-244636
#>   [881]     chrY 16842593-16846893      * |          NLGN4Y         208105
#>   [882]     chrY 16998750-17003225      * |          NLGN4Y         364262
#>   [883]     chrY 24401612-24406252      * |   RBMY1E,RBMY1F  -72523,-48754
#>   [884]     chrY 26206997-26215445      * |            DAZ2         841416
#>   -------
#>   seqinfo: 24 sequences from an unspecified genome; no seqlengths