Get a single enrichment table from GREAT web server

Get a single enrichment table from GREAT web server

# S4 method for GreatJob
getEnrichmentTable(object, ontology, ...)

Arguments

object

A GreatJob-class object returned by submitGreatJob.

ontology

A single ontology names. Valid values are in availableOntologies.

...

All pass to getEnrichmentTables,GreatJob-method.

Value

A data frame of the enrichment results for a single ontology.

Examples

# note the `job` was generated from GREAT 3.0.0
job = readRDS(system.file("extdata", "GreatJob.rds", package = "rGREAT"))
tb = getEnrichmentTable(job, ontology = "GO Molecular Function")
#> The default enrichment tables contain no associated genes for the input
#> regions.You can set `download_by = 'tsv'` to download the complete
#> table,but note only the top 500 regions can be retreived. See the
#> following link:
#> 
#> https://great-help.atlassian.net/wiki/spaces/GREAT/pages/655401/Export#Export-GlobalExport
head(tb)
#>           ID                            name Binom_Genome_Fraction
#> 1 GO:0005515                 protein binding             0.7515051
#> 2 GO:0005488                         binding             0.8694407
#> 3 GO:0097159 organic cyclic compound binding             0.4561508
#> 4 GO:1901363   heterocyclic compound binding             0.4511336
#> 5 GO:0003723                     RNA binding             0.1199843
#> 6 GO:0003676            nucleic acid binding             0.3173138
#>   Binom_Expected Binom_Observed_Region_Hits Binom_Fold_Enrichment
#> 1      2816.6410                       3186              1.131135
#> 2      3258.6640                       3505              1.075594
#> 3      1709.6530                       2050              1.199074
#> 4      1690.8490                       2031              1.201172
#> 5       449.7013                        669              1.487654
#> 6      1189.2920                       1482              1.246119
#>   Binom_Region_Set_Coverage Binom_Raw_PValue Binom_Adjp_BH Hyper_Total_Genes
#> 1                 0.8500534     2.435386e-49  1.027489e-45             11129
#> 2                 0.9351654     4.308462e-39  9.088701e-36             14268
#> 3                 0.5469584     5.192796e-29  5.646545e-26              5936
#> 4                 0.5418890     5.353444e-29  5.646545e-26              5853
#> 5                 0.1784952     1.794959e-25  1.514586e-22              1628
#> 6                 0.3954109     3.948077e-24  2.776156e-21              3991
#>   Hyper_Expected Hyper_Observed_Gene_Hits Hyper_Fold_Enrichment
#> 1      2302.7170                     2635              1.144300
#> 2      2952.2120                     3159              1.070045
#> 3      1228.2260                     1417              1.153696
#> 4      1211.0530                     1400              1.156019
#> 5       336.8518                      448              1.329962
#> 6       825.7835                     1005              1.217026
#>   Hyper_Gene_Set_Coverage Hyper_Term_Gene_Coverage Hyper_Raw_PValue
#> 1               0.6865555                0.2367688     1.078774e-35
#> 2               0.8230849                0.2214045     5.578685e-20
#> 3               0.3692027                0.2387129     2.063283e-13
#> 4               0.3647733                0.2391936     1.630373e-13
#> 5               0.1167275                0.2751843     3.226236e-12
#> 6               0.2618551                0.2518166     4.628398e-15
#>   Hyper_Adjp_BH
#> 1  4.551348e-32
#> 2  1.176824e-16
#> 3  1.740998e-10
#> 4  1.719636e-10
#> 5  1.723691e-09
#> 6  6.509070e-12