Get region-gene associations

Get region-gene associations

# S4 method for GreatJob
getRegionGeneAssociations(object, ontology = NULL, term_id = NULL,
    request_interval = 10, max_tries = 100, verbose = great_opt$verbose)

Arguments

object

A GreatJob-class object returned by submitGreatJob.

ontology

ontology name

term_id

Term id in the selected ontology.

request_interval

Time interval for two requests. Default is 300 seconds.

max_tries

Maximal times for automatically reconnecting GREAT web server.

verbose

Whether to show messages.

Value

A GRanges object. Please the two meta columns are in formats of CharacterList

and IntegerList because a region may associate to multiple genes.

Please note, the distance is from the middle points of input regions to TSS.

Author

Zuguang gu <z.gu@dkfz.de>

Examples

job = readRDS(system.file("extdata", "GreatJob.rds", package = "rGREAT"))

gr = getRegionGeneAssociations(job)
gr
#> GRanges object with 3734 ranges and 2 metadata columns:
#>          seqnames              ranges strand |        annotated_genes
#>             <Rle>           <IRanges>  <Rle> |        <CharacterList>
#>      [1]     chr1     1310672-1311002      * |               AURKAIP1
#>      [2]     chr1     1342595-1342925      * |                 MRPL20
#>      [3]     chr1     1509773-1510103      * |                  SSU72
#>      [4]     chr1     1711637-1711967      * |                   NADK
#>      [5]     chr1     2323001-2323331      * |                   RER1
#>      ...      ...                 ...    ... .                    ...
#>   [3730]     chrX 130912591-130912921      * | ENSG00000134602,OR13H1
#>   [3731]     chrX 130947194-130947524      * | ENSG00000134602,OR13H1
#>   [3732]     chrX 134449748-134450078      * |            CT55,ZNF75D
#>   [3733]     chrX 138877458-138877788      * |            MCF2,ATP11C
#>   [3734]     chrX 147691650-147691980      * |               AFF2,IDS
#>             dist_to_TSS
#>           <IntegerList>
#>      [1]           -300
#>      [2]            -67
#>      [3]            311
#>      [4]          -1893
#>      [5]           -106
#>      ...            ...
#>   [3730] -244537,234806
#>   [3731] -209934,269409
#>   [3732]  -144591,28099
#>   [3733]   -87237,36824
#>   [3734]  109676,895062
#>   -------
#>   seqinfo: 23 sequences from an unspecified genome; no seqlengths